+++ test/unit/File/FileTest.php
@@ -80,7 +80,7 @@ class FileTest extends TestCase
      */
     public function testCreateFromFileWithInvalidId()
     {
-        $path = __DIR__ . '/../../data/yml/invalid-pipeline-id.yml';
+        $path = __DIR__ . '/../../data/yml/invalid/pipeline-id.yml';
 
         $file = File::createFromFile($path);